move cell classes to separate files
